src/backend/pyproject.toml
[project]
name = "eduscale-engine"
version = "1.0.0"
description = "EduScale Engine - Open-source data infrastructure for Eduzměna Foundation. DigiEduHack 2025 solution for scaling educational impact across 5,000+ Czech schools."
requires-python = ">=3.10,<4.0"

dependencies = [
    # FastAPI & ASGI
    "fastapi[standard]<1.0.0,>=0.114.2",
    "uvicorn[standard]<1.0.0,>=0.23.0",
    "gunicorn<22.0.0,>=20.1.0",
    # Database and ORM
    "sqlalchemy<3.0,>=2.0",
    "sqlmodel<1.0.0,>=0.0.21",
    "alembic<2.0.0,>=1.12.1",
    "asyncpg",
    "psycopg[binary]<4.0.0,>=3.1.13",
    # Pydantic
    "pydantic>=2.11.7",
    "pydantic-core>=2.33.2",
    "pydantic-settings<3.0.0,>=2.2.1",
    # Auth & Security
    "passlib[bcrypt]<2.0.0,>=1.7.4",
    "bcrypt==4.0.1",
    "pyjwt<3.0.0,>=2.8.0",
    "authlib>=1.4.0",
    "itsdangerous>=2.2.0",
    # Email validation
    "email-validator<3.0.0.0,>=2.1.0.post1",
    # HTTP & SSE
    "httpx<1.0.0,>=0.25.1",
    "httpx-sse==0.4.1",
    "openai>=1.37.0",
    # Templating & Utils
    "jinja2<4.0.0,>=3.1.4",
    "python-multipart<1.0.0,>=0.0.7",
    "psutil",
    "tenacity<9.0.0,>=8.2.3",
    "openapi-pydantic==0.5.1",
    # ngrok for tunneling
    "ngrok>=1.4.0",
    # Sentry
    "sentry-sdk[fastapi]",
    ##caching
    "async-lru==2.0.5",
    #other
    "rapidfuzz==3.13.0",
    "dateparser",
    "redis[async]>=4.5.1",
    "czechfabric-sdk>=0.2.9",
    # TONE format support
    "toneformat>=1.0.0",
    # Data ingestion & processing
    "pandas>=2.2.0",
    "openpyxl>=3.1.2",
    "python-docx>=1.0.0",
    # Documentation generation
    "reportlab>=4.2.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.0,<9.0.0",
    "pytest-asyncio==1.0.0",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "black<24.0.0,>=23.12.0",
    "pre-commit<4.0.0,>=3.6.2",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
]

[build-system]
requires = ["hatchling", "uv"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "swx_api",
    "czfb_server"
]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "migrations", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["migrations", "alembic"]

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # Pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG001", # unused arguments
]
ignore = [
    "E501", # line length
    "B008", # function calls in argument defaults
    "W191", # tabs
    "B904", # exception chaining
]

[project.scripts]
swx = "swx_api.core.cli.main:main"
